requirements.txt
requests
beautifulsoup4
prefect
sqlalchemy
psycopg2-binary
pandas
python-dotenv
